cleaned up existing stats functions
Made several changes to `lipydomics/stats.py` module to clean up the stats functions. First, took the label kwarg out of both functions, since it is more reliable to label these stats entries in a consistent and systematic way. Second, refactored the PCA loadings function so that it automatically adds both the loadings and the projections from a 3 component PCA to the Dataset. Changed the name to `add_ pca3` and got rid of the unimplemented `get_pca3_projections` function. Updated documentation to reflect.