Changed how PCA3 works (Fixes #3)
(Fixes #3) Changed `add_pca3(...)` so that it takes group names as a parameter, that way the user may choose a subset of the data to include when computing the PCA. As a caveat, when going to plot the projections, the user must use the same groups (in the same order) to select the proper `Dataset.stats` entry. Updated documentation accordingly.