changed behavior of interactive barplot feature bygroup (Fixes #21 )
Changed the plotting function barplot_feature_bygroup so that it returns a boolean indicating whether a feature was actually found. Changed the interactive plotting function to handle this change in behavior accordingly. Also modified the try/except clauses to explicitly handle a ValueError from the plotting function and report that to the user before printing its own error message.