had to edit the lipid parser to recognize PIP2 and PIP3
building the lipids.db database was failing because the lipid parser was failing to parse PIP2 and PIP3 lipid names. This was happening because the regex for the lipid class would not recognize any numbers... added 2 and 3 to that part of the regex. Rebuilt the database and everything looks good now.