small fixes to custom ID levels and testing overhaul
Made a couple small fixes on the code for doing custom ID levels.
Overhauled the testing code so that tests can either be run at a global level (`python3 -m lipydomics.test`) or at a module level (`python3 -m lipydomics.test.identification`). Now the lists of tests to run are defined in the individual modules and imported by `test/__main__.py` so when new tests are added, only the individual module needs to be modified not both.