implemented the public `drop_features` method, added a couple tests
Implemented the logic in the public method, includes input validation and constructing the target boolean array and passing that along to the (private) helper function. Also added a couple of test functions that make sure input validation is working and that the actual drop_features method is working. Also updated html documentation